Moved CampaignStart and Settings pages into folders.
