+++ src/App.jsx
@@ -12,14 +12,14 @@ import WeVoteRouter from './js/components/Widgets/WeVoteRouter';
 const About = React.lazy(() => import('./js/pages/About'));
 const Attributions = React.lazy(() => import('./js/pages/Attributions'));
 const CampaignDetailsPage = React.lazy(() => import('./js/pages/CampaignDetailsPage'));
-const CampaignStartAddPolitician = React.lazy(() => import('./js/pages/CampaignStartAddPolitician'));
-const CampaignStartAddDescription = React.lazy(() => import('./js/pages/CampaignStartAddDescription'));
-const CampaignStartAddPhoto = React.lazy(() => import('./js/pages/CampaignStartAddPhoto'));
-const CampaignStartAddTitle = React.lazy(() => import('./js/pages/CampaignStartAddTitle'));
+const CampaignStartAddPolitician = React.lazy(() => import('./js/pages/CampaignStart/CampaignStartAddPolitician'));
+const CampaignStartAddDescription = React.lazy(() => import('./js/pages/CampaignStart/CampaignStartAddDescription'));
+const CampaignStartAddPhoto = React.lazy(() => import('./js/pages/CampaignStart/CampaignStartAddPhoto'));
+const CampaignStartAddTitle = React.lazy(() => import('./js/pages/CampaignStart/CampaignStartAddTitle'));
 const CampaignStartCompleteYourProfileMobile = React.lazy(() => import('./js/pages/CampaignStart/CampaignStartCompleteYourProfileMobile'));
-const CampaignStartEditAll = React.lazy(() => import('./js/pages/CampaignStartEditAll'));
-const CampaignStartIntro = React.lazy(() => import('./js/pages/CampaignStartIntro'));
-const CampaignStartPreview = React.lazy(() => import('./js/pages/CampaignStartPreview'));
+const CampaignStartEditAll = React.lazy(() => import('./js/pages/CampaignStart/CampaignStartEditAll'));
+const CampaignStartIntro = React.lazy(() => import('./js/pages/CampaignStart/CampaignStartIntro'));
+const CampaignStartPreview = React.lazy(() => import('./js/pages/CampaignStart/CampaignStartPreview'));
 const Credits = React.lazy(() => import('./js/pages/Credits'));
 const FAQ = React.lazy(() => import('./js/pages/FAQ'));
 const HomePage = React.lazy(() => import('./js/pages/HomePage'));
@@ -28,7 +28,7 @@ const Membership = React.lazy(() => import('./js/pages/Membership'));
 const PageNotFound = React.lazy(() => import('./js/pages/PageNotFound'));
 const Privacy = React.lazy(() => import('./js/pages/Privacy'));
 const SettingsEditProfile = React.lazy(() => import('./js/pages/Settings/SettingsEditProfile'));
-const SettingsYourCampaigns = React.lazy(() => import('./js/pages/SettingsYourCampaigns'));
+const SettingsYourCampaigns = React.lazy(() => import('./js/pages/Settings/SettingsYourCampaigns'));
 const StyleGuidePage = React.lazy(() => import('./js/pages/StyleGuidePage'));
 const TermsOfService = React.lazy(() => import('./js/pages/TermsOfService'));
 const TwitterSignInProcess = React.lazy(() => import('./js/pages/TwitterSignInProcess'));
