+++ src/js/pages/CampaignStart/CampaignStartAddPolitician.jsx
@@ -4,15 +4,15 @@ import { withStyles } from '@material-ui/core/styles';
 import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 import styled from 'styled-components';
-import CampaignStartActions from '../actions/CampaignStartActions';
-import AddCandidateInputField from '../components/CampaignStart/AddPoliticianInputField';
-import CampaignStartSteps from '../components/Navigation/CampaignStartSteps';
-import MainFooter from '../components/Navigation/MainFooter';
-import MainHeaderBar from '../components/Navigation/MainHeaderBar';
-import CampaignStartStore from '../stores/CampaignStartStore';
-import { historyPush, isCordova } from '../utils/cordovaUtils';
-import initializejQuery from '../utils/initializejQuery';
-import { renderLog } from '../utils/logging';
+import CampaignStartActions from '../../actions/CampaignStartActions';
+import AddCandidateInputField from '../../components/CampaignStart/AddPoliticianInputField';
+import CampaignStartSteps from '../../components/Navigation/CampaignStartSteps';
+import MainFooter from '../../components/Navigation/MainFooter';
+import MainHeaderBar from '../../components/Navigation/MainHeaderBar';
+import CampaignStartStore from '../../stores/CampaignStartStore';
+import { historyPush, isCordova } from '../../utils/cordovaUtils';
+import initializejQuery from '../../utils/initializejQuery';
+import { renderLog } from '../../utils/logging';
 
 class CampaignStartAddPolitician extends Component {
   constructor (props) {
